requirements.txt
# ========================================
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# WSGI Server for Production
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# AI API (DeepSeek Compatible)
openai==1.12.0

# Database (PostgreSQL Driver)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Payment SDKs (可选，根据实际需要)
# alipay-sdk-python==3.0.382
# wechatpayv2==1.8.1
